Python creation of candidates and targets
