+++ src/data/module_logger_config.json
@@ -29,7 +29,7 @@
                 "stream_handler"
             ],
             "level": "DEBUG",
-            "propagate": true
+            "propagate": false
         }
     }
 }
